+++ App/app/src/main/java/com/example/carworth/uix/view/fonk/Araba.kt
@@ -25,7 +25,7 @@ data class Araba(
     val seri: String,
     val model: String,
     val yil: Int,
-    val kilometre: Int,
+    val kilometre: Double,
     val vitesTipi: String,
     val yakitTipi: String,
     val kasaTipi: String,
